+++ Editor/ValidationSuite/ValidationTests/Standards/US0017-CorrectPackageVersionTags.cs
@@ -18,12 +18,6 @@ namespace UnityEditor.PackageManager.ValidationSuite.ValidationTests.Standards
                 return;
             }
 
-            if (packageVersionNumber.Major < 1) //TODO: this is not captured by a standard
-            {
-                AddError("In package.json, \"version\" cannot be tagged \"" + packageVersionNumber.Prerelease + "\" while the major version less than 1. " + ErrorDocumentation.GetLinkMessage(ErrorTypes.InvalidLifecycleV2));
-                return;
-            }
-
             if (versionTag.Tag != "exp" && versionTag.Tag != "pre")
             {
                 AddError("In package.json, \"version\" must be a valid tag. \"" + versionTag.Tag + "\" is invalid, try either \"pre\" or \"exp\". " + ErrorDocumentation.GetLinkMessage(ErrorTypes.InvalidLifecycleV2));